+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -259,6 +259,30 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     }
 
     @Test
+    @TestMetadata("defaultParameters.kt")
+    public void testDefaultParameters() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/defaultParameters.kt");
+    }
+
+    @Test
+    @TestMetadata("exposeImplBinding.kt")
+    public void testExposeImplBinding() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/exposeImplBinding.kt");
+    }
+
+    @Test
+    @TestMetadata("implTypeNotDirectlyAccessible.kt")
+    public void testImplTypeNotDirectlyAccessible() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/implTypeNotDirectlyAccessible.kt");
+    }
+
+    @Test
+    @TestMetadata("implicitClassKeyMapKey.kt")
+    public void testImplicitClassKeyMapKey() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/implicitClassKeyMapKey.kt");
+    }
+
+    @Test
     @TestMetadata("intoMap.kt")
     public void testIntoMap() {
       runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/intoMap.kt");
@@ -286,6 +310,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     @TestMetadata("providerAndLazy.kt")
     public void testProviderAndLazy() {
       runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/providerAndLazy.kt");
+    }
+
+    @Test
+    @TestMetadata("qualifiedBinding.kt")
+    public void testQualifiedBinding() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/qualifiedBinding.kt");
     }
 
     @Test